% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.rules}
\alias{as.rules}
\title{Converts a vector of lines into a rules data frame}
\usage{
as.rules(lines = "")
}
\arguments{
\item{lines}{R statements with conditions}
}
\value{
The results as a datadict object or NA for 'empty' rule lines
}
\description{
The rules must be one per line and should evaluate to a vector of TRUE or FALSE.
}
\details{
A rule must only refer to one 'column' name.
Rule statements may not have assignment operators (= or <-).
Rules may be separated by empty or commented lines.
A comment after a rule is used to document the specific rule in the summary table.
}
\author{
Reinhard Simon
}
\seealso{
Other datadict: \code{\link{datadict.profile}};
  \code{\link{has.ruleErrors}};
  \code{\link{is.datadict.profile}};
  \code{\link{prep4rep}}; \code{\link{read.rules}}
}

