% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{checker}
\alias{checker}
\title{Check a dataframe for presence of a regular expression.}
\usage{
checker(df, regex)
}
\arguments{
\item{df}{Data frame.}

\item{regex}{Regular expression in string object.}
}
\value{
Boolean.
}
\description{
Check a dataframe for presence of a regular expression.
}
