% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck-package.R
\docType{package}
\name{datacheck-package}
\alias{datacheck-package}
\alias{_PACKAGE}
\title{datacheck: Check Data For Potential Privacy Violations}
\description{
The datacheck package scans datasets for potential privacy violations according to data protection laws (e.g., GDPR, HIPAA). This is a tool to help screen datasets that researchers may want to openly share online. It can also be used as a research tool to scan datasets for potential privacy violations. The data is not sent to any remote server for this screening, it happens 100% with the package and on the person's device.
}
\author{
\strong{Maintainer}: Chris Hartgerink \email{chris@libscie.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID})

Authors:
\itemize{
  \item Richard Klein \email{raklein22@gmail.com} (\href{https://orcid.org/0000-0002-8713-1972}{ORCID})
}

}
\keyword{internal}
