% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{datacheck}
\alias{datacheck}
\title{Check a path for identifying information.}
\usage{
datacheck(path, folder = FALSE)
}
\arguments{
\item{path}{A system path (file or folder).}

\item{folder}{Boolean, indicate whether the path is a folder (defaults to FALSE).}
}
\value{
Data frame, rows (files), columns (Boolean; presence of identifier).
}
\description{
Check a path for identifying information.
}
