% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheckDf.R
\name{datacheckDf}
\alias{datacheckDf}
\title{Check a data file for identifying information.}
\usage{
datacheckDf(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Data frame, with columns for each type of identifier (Boolean).
}
\description{
Check a data file for identifying information.
}
