% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{datacheckFile}
\alias{datacheckFile}
\title{Check a data file for identifying information.}
\usage{
datacheckFile(path)
}
\arguments{
\item{path}{System path (file only).}
}
\value{
Data frame, with columns for each type of identifier (Boolean).
}
\description{
Check a data file for identifying information.
}
