% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{datacheckFolder}
\alias{datacheckFolder}
\title{Check a folder for identifying information.}
\usage{
datacheckFolder(path)
}
\arguments{
\item{path}{System path (folder only).}
}
\value{
Data frame, rows (files), columns (Boolean; presence of identifier).
}
\description{
Check a folder for identifying information.
}
