% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{has.punct}
\alias{has.punct}
\title{Tests for presence of most common punctuation characters}
\usage{
has.punct(s)
}
\arguments{
\item{s}{a character string}
}
\value{
boolean TRUE if detects anything
}
\description{
These include: .?'!_#$^&*()+=<>,;' - and [tab] and [at]
}
\examples{

has_punct(".") == TRUE 
}
\author{
Reinhard Simon
}
\seealso{
Other rule_checks: \code{\link{has_punct}};
  \code{\link{is.oneOf}}; \code{\link{is.onlyLowers}};
  \code{\link{is.properName}};
  \code{\link{is.withinRange}}; \code{\link{is_one_of}};
  \code{\link{is_only_lowers}};
  \code{\link{is_proper_name}};
  \code{\link{is_within_range}}
}

