% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{heatmap_quality}
\alias{heatmap_quality}
\title{Draws a heatmap based on data quality scores}
\usage{
heatmap_quality(profile, recLab = NULL, recMax = 100, scoreMax = NULL,
  cols = NULL, ...)
}
\arguments{
\item{profile}{a datadict.profile object}

\item{recLab}{variable that should be used for labeling the records}

\item{recMax}{maximum first n records for display}

\item{scoreMax}{maxim quality score to filter out}

\item{cols}{color scheme}

\item{...}{as in heatmap function}
}
\description{
Knows to extract the quality matrix from the profile object and pass it on to the heatmap function.
Plots a heatmap.
}
\details{
Currently this function is limited a table siZé of 300 records.
}
\author{
Reinhard Simon
}
\seealso{
Other visuals: \code{\link{heatmap.quality}};
  \code{\link{ruleCoverage}}; \code{\link{rule_coverage}};
  \code{\link{scoreSum}}; \code{\link{score_sum}}
}

