% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{is_proper_name}
\alias{is_proper_name}
\title{Tests if string is like a proper name with inital letter in upper case}
\usage{
is_proper_name(aname)
}
\arguments{
\item{aname}{a character string}
}
\value{
boolean TRUE if ok
}
\description{
Tests if string is like a proper name with inital letter in upper case
}
\examples{
# Valid proper names

is_proper_name("John") == TRUE

is_proper_name("john") == FALSE

is_proper_name(123) == FALSE 
}
\author{
Reinhard Simon, Jose Francisco Loff
}
\seealso{
Other rule_checks: \code{\link{has.punct}};
  \code{\link{has_punct}}; \code{\link{is.oneOf}};
  \code{\link{is.onlyLowers}}; \code{\link{is.properName}};
  \code{\link{is.withinRange}}; \code{\link{is_one_of}};
  \code{\link{is_only_lowers}};
  \code{\link{is_within_range}}
}

