% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{is_within_range}
\alias{is_within_range}
\title{Tests if a numeric value is between a minimal and maximum value. Serves as convenience function.}
\usage{
is_within_range(val, min, max)
}
\arguments{
\item{val}{The value to be checked}

\item{min}{The minimal value (inclusive)}

\item{max}{The maximum value (inclusive)}
}
\value{
boolean TRUE if detects anything
}
\description{
Tests if a numeric value is between a minimal and maximum value. Serves as convenience function.
}
\examples{
is_within_range(1, 0, 2) == TRUE

is_within_range(-1, 0, 2) == FALSE 
}
\author{
Reinhard Simon
}
\seealso{
Other rule_checks: \code{\link{has.punct}};
  \code{\link{has_punct}}; \code{\link{is.oneOf}};
  \code{\link{is.onlyLowers}}; \code{\link{is.properName}};
  \code{\link{is.withinRange}}; \code{\link{is_one_of}};
  \code{\link{is_only_lowers}};
  \code{\link{is_proper_name}}
}

