% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{pkg_version}
\alias{pkg_version}
\title{Get the current version of a package}
\usage{
pkg_version(pkg)
}
\arguments{
\item{pkg}{the package name}
}
\value{
a string with the package number
}
\description{
Uses the citation() function.
}
\author{
Reinhard Simon
}
\seealso{
Other helper: \code{\link{pkg.version}};
  \code{\link{shortSummary}}; \code{\link{short_summary}}
}

