% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{run_datacheck}
\alias{run_datacheck}
\title{Presents the packages graphical user interface}
\usage{
run_datacheck(port = 1971L)
}
\arguments{
\item{port}{the port where to listen; 1971 by default.}
}
\description{
Runs a web server to show the user interface.
}
\author{
Reinhard Simon
}
\seealso{
Other interface: \code{\link{runDatacheck}}
}

