% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runDatacheck}
\alias{runDatacheck}
\title{Presents the packages graphical user interface}
\usage{
runDatacheck(port = 1971L)
}
\arguments{
\item{port}{the port where to listen; 1971 by default.}
}
\description{
Runs a web server to show the user interface.
}
\author{
Reinhard Simon
}

