% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{score_sum}
\alias{score_sum}
\title{Line chart of cumulative sum of rule scores.}
\usage{
score_sum(profile)
}
\arguments{
\item{profile}{a datadict profile object}
}
\description{
A record receives one point per rule which evaluates TRUE.
The total number of points is the 'quality score' per record.
}
\author{
Reinhard Simon
}
\seealso{
Other visuals: \code{\link{heatmap.quality}};
  \code{\link{heatmap_quality}};
  \code{\link{ruleCoverage}}; \code{\link{rule_coverage}};
  \code{\link{scoreSum}}
}

