% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacheck.R
\name{shortSummary}
\alias{shortSummary}
\title{Produces a tabular summary of descriptive statistics using the 'describe'
function from the Hmisc package.}
\usage{
shortSummary(atable)
}
\arguments{
\item{atable}{a data frame}
}
\description{
Returns a dataframe of descriptive statistics
}
\author{
Reinhard Simon
}
\seealso{
Other helper: \code{\link{pkg.version}}
}

